+++ solana/programs/revenue-distribution/CHANGELOG.md
@@ -3,10 +3,12 @@
 ## Unreleased
 
 - add null rewards root protection ([#86])
+- fix swap balance in journal ([#87])
 
 ## [v0.1.0] 
 
 - Initial release
 
 [#86]: https://github.com/doublezerofoundation/doublezero-solana/pull/86
+[#87]: https://github.com/doublezerofoundation/doublezero-solana/pull/87
 [v0.1.0]: https://github.com/doublezerofoundation/doublezero-solana/tree/revenue-distribution/v0.1.0